[✨ feat] Service 인터페이스 분리
- 확장성을 고려하여, 서비스단을 인터페이스와 구현체로 분리했습니다.
